webpack.config.jsとwebpack.config.base.jsに分離
